requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
line-bot-sdk>=3.11.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
aiohttp>=3.9.0
